[DotNetCore] Add .NET Core runtime "component"
